Add get_graph_from_model_generative_test
